refactor: move `escape_special_characters` function to the util file
